requirements.txt
Django
djangorestframework
psycopg2
gunicorn
Pillow
requests
django-extensions
django_storages
boto3
pyyaml
